+++ examples/benchmarks/smax_benchmarks/smax_sptrsv_lvl.cpp
@@ -42,7 +42,7 @@ int main(int argc, char *argv[]) {
 
     // Make lambda, and pass to the benchmarking harness
     std::string bench_name = "smax_sptrsv_lvl";
-    double runtime = 0.0;
+    float runtime = 0.0;
     int n_iter = MIN_NUM_ITERS;
     int n_threads = 1;
 #ifdef _OPENMP
